fix(#1752): close the call-site gaps the round-2 review found
- tandem_edit refuses a point insertion that carries a textSnapshot. A
  zero-length range matches no text, so the snapshot always reads stale and the
  RANGE_MOVED retry would silently turn the insert into a replacement.
- The watcher's relocation PROBE now passes `surrogates: "ignore"` and shares
  that binding with the anchor call. The probe is a capped slice of a STORED
  snapshot, so a cap landing mid-pair was reporting a relocation failure on a
  document that had not changed.
- Both watcher calls and the whole-doc authorship stamp carry a HoistTag.
- populate.ts gains an exhaustiveness `never` on the LoadIssue switch, so a new
  variant that forgets its notification is a type error rather than silence, and
  exports notifyIssue for the spec that pins the clamp notification.
- The three deprecated annotation stubs get `.int()` on from/to. This narrows
  their "DEPRECATED for any call shape" invariant by exactly one case — a
  fractional offset now answers a protocol error instead — recorded in place.
- docx-comment-export: the outward snap comment says what is actually proven
  (a single step is total) and no longer claims the invalid-range arm is a
  second layer; it is there for out-of-bounds/inverted/non-integer.
- The whole-doc stamp's skip arm records why no spec pins it: block ranges start
  after a heading prefix and end at a block boundary, so no failure reason is
  constructible. The log is for the day that stops being true.

+++ src/server/mcp/annotations.ts
@@ -233,8 +233,15 @@ export function registerAnnotationTools(server: McpServer): void {
     {
       // All params optional: a deprecated stub must surface DEPRECATED for any
       // call shape, including ones missing the legacy required params.
-      from: z.number().optional(),
-      to: z.number().optional(),
+      //
+      // `.int()` matches the three LIVE range-taking tools (#1752). It narrows
+      // "any call shape" by exactly one case — a fractional offset now answers a
+      // schema error instead of DEPRECATED — and that is the intended trade: the
+      // rule about offsets should not have a hole in it just because the tool
+      // behind it is a stub, and no legacy caller passed a fraction. Optionality,
+      // which is what the invariant is actually about, is untouched.
+      from: z.number().int().optional(),
+      to: z.number().int().optional(),
       color: HighlightColorSchema.optional(),
       note: z.string().optional(),
       documentId: z.string().optional(),
@@ -333,8 +340,15 @@ export function registerAnnotationTools(server: McpServer): void {
     {
       // All params optional: a deprecated stub must surface DEPRECATED for any
       // call shape, including ones missing the legacy required params.
-      from: z.number().optional(),
-      to: z.number().optional(),
+      //
+      // `.int()` matches the three LIVE range-taking tools (#1752). It narrows
+      // "any call shape" by exactly one case — a fractional offset now answers a
+      // schema error instead of DEPRECATED — and that is the intended trade: the
+      // rule about offsets should not have a hole in it just because the tool
+      // behind it is a stub, and no legacy caller passed a fraction. Optionality,
+      // which is what the invariant is actually about, is untouched.
+      from: z.number().int().optional(),
+      to: z.number().int().optional(),
       newText: z.string().optional(),
       reason: z.string().optional(),
       documentId: z.string().optional(),
@@ -355,8 +369,15 @@ export function registerAnnotationTools(server: McpServer): void {
     {
       // All params optional: a deprecated stub must surface DEPRECATED for any
       // call shape, including ones missing the legacy required params.
-      from: z.number().optional(),
-      to: z.number().optional(),
+      //
+      // `.int()` matches the three LIVE range-taking tools (#1752). It narrows
+      // "any call shape" by exactly one case — a fractional offset now answers a
+      // schema error instead of DEPRECATED — and that is the intended trade: the
+      // rule about offsets should not have a hole in it just because the tool
+      // behind it is a stub, and no legacy caller passed a fraction. Optionality,
+      // which is what the invariant is actually about, is untouched.
+      from: z.number().int().optional(),
+      to: z.number().int().optional(),
       note: z.string().optional(),
       documentId: z.string().optional(),
       textSnapshot: z.string().optional(),
